+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 4.0.2-wip
+## 5.0.0-wip
 
-- Update to the latest version of `package:dart_flutter_team_lints`.
+- Update to the latest version of `package:dart_flutter_team_lints`
+- Using internal futures list to store send events
 
 ## 4.0.1